Fix some thread-safety issues in SpiderMonkeyUtils.cpp, and add lots of explanatory comments.
I was using lambdas as IIFEs to initialize thread-local statics, as a sort of poor-mans one-time function. However,
I was forgetting that while the standard allows initialization of such statics to be left until first execution -
which most compilers do, a behaviour I was relying on - it is still possible that they might be initialized earlier,
breaking a lot of my assumptions.

Additionally, fixes a leak where we never destroyed a TLS key.